refactor: :truck: クラス名, ファイル名, ディレクトリ構成を変更
